+++ src/mocks/mockIStorageService.ts
@@ -28,13 +28,7 @@
 import powerbi from "powerbi-visuals-api";
 import IPromise = powerbi.IPromise;
 import PrivilegeStatus = powerbi.PrivilegeStatus;
-
-export interface ILocalVisualStorageService {
-    status(): IPromise<PrivilegeStatus>
-    get(key: string): IPromise<string>;
-    set(key: string, data: string): IPromise<number>;
-    remove(key: string): void;
-}
+import ILocalVisualStorageService = powerbi.extensibility.ILocalVisualStorageService
 
 function getLocalStorageStatus() {
     try {